DatabaseD replacement for Database class prepared.
MainApp.java purified of redundant code, added methods to display logon dialog. Code is now more readable.

Model Goods.java updated with setters and getters to reflect reality. Most of the data types are still not consistent with those stored in database.

Credentials file now contains comment on how to work with it. This can be now considered final version of the file.